+++ src/utils/index.ts
@@ -6,6 +6,7 @@ import generateUUID from "./generateUUID";
 import getMostContrastingColor from "./getMostContrastingColor";
 import getMostDifferentHue from "./getMostDifferentHue";
 import getRandomFromSlice from "./getRandomFromSlice";
+import { trackEvent, initAnalytics } from "./analytics";
 
 export {
   calculateAspectRatioDimensions,
@@ -16,4 +17,6 @@ export {
   getMostContrastingColor,
   getMostDifferentHue,
   getRandomFromSlice,
+  trackEvent,
+  initAnalytics,
 };